requirements.txt
diffusers==0.16.1
importlib-resources==5.10.2
matplotlib==3.6.3
numpy==1.24.2
Pillow==9.4.0
ruamel.yaml==0.17.21
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
torch==1.13.0+cu116
torchvision==0.14.0+cu116
tqdm==4.64.1
transformers==4.27.4